requirements.txt
autopep8==1.4.4
certifi==2019.3.9
Django==2.2.2
djangorestframework==3.9.4
Markdown==3.1.1
pycodestyle==2.5.0
pytz==2019.1
sqlparse==0.3.0
wincertstore==0.2